requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
PyJWT==2.6.0
bcrypt==4.0.1
openai==1.52.2
google-generative-ai==0.4.0
python-dotenv==1.0.0
